+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Removed RcVecUtils.Subtract(float[] verts, int i, int j)
 - Removed RcMeshDetails.VdistSq2(float[], float[])
 - Removed RcVecUtils.Min(), RcVecUtils.Max()
+- Removed RcVecUtils.Create(float[] values)
 
 ### Special Thanks
 - [@Doprez](https://github.com/Doprez)